feat(parser): added and updated some parser modules
